BugFix: NextpeerNotifier queueScheduler should accept no param; Add Nextpeer getClass JNI as Cocos2d-x does not support that any more

+++ Nextpeer-Cocos2D-X/Classes/NextpeerNotifier.cpp
@@ -26,7 +26,7 @@ namespace nextpeer
     void NextpeerNotifier::queueScheduler(float dt)
     {
         NextpeerEventQueue* queue = NextpeerEventQueue::getInstance();
-        unsigned int count = queue->queueSize();
+        unsigned long count = queue->queueSize();
 
         if (!count) return;
